requirements.txt
Django==4.0.4
environs==9.5.0
django-debug-toolbar==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
stripe==2.74.0
django-cors-headers==3.11.0
